requirements.txt
streamlit
Pillow
numpy
scikit-learn
matplotlib